Move the crontab decorator to a better place
